feat: modularize deploy
