+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "tidebit-defi",
-  "version": "0.9.0+114",
+  "version": "0.9.0+115",
   "private": true,
   "type": "module",
   "scripts": {
